separate types into different file reorganise imports update README
